com.buglabs.bug.bmi, com.buglabs.bug.emulator.bmi - Added context.removeServiceListener(this) to both the vbug and real bug activators' stop() method.
For some reason the new version of concierge was calling service changed events on the bmi bundle activator on shut-down, after the bmi bundle had been stopped, and it was causing a stream of errors when shutting down the virtual bug.  (Actually, calling the service event seems to be the correct behavior, so it seems it was the old version that was broken).  The fix is in the BMI activator code, on stop(), to remove itself as a serviceListener.


git-svn-id: svn://bugcamp.net/bug/trunk/com.buglabs.bug.bmi@9630 35ad6e01-222d-0410-b96a-b41bba574936